Update README.md
This program is Monte Carlo simulation to determine the probability that two people with have the same birthdays, even in a small group of people. The simulation performs several probabilistic experiments to determine the percentages for groups of different sizes. We are running multiple random trials to understand the likely outcomes of each simulation. Running 100,000 simulations can take quite a bit of time, which is why the program updates the user after every 10,000 simulations. This feedback is to assure the user that the program has not stopped working.